+++ src/main/java/com/ynyes/lyz/entity/TdSmsAccount.java
@@ -37,7 +37,7 @@ public class TdSmsAccount {
 
 	// 排序号
 	@Column
-	private Long sortId;
+	private Double sortId;
 
 	public Long getId() {
 		return id;
@@ -79,11 +79,11 @@ public class TdSmsAccount {
 		this.userName = userName;
 	}
 
-	public Long getSortId() {
+	public Double getSortId() {
 		return sortId;
 	}
 
-	public void setSortId(Long sortId) {
+	public void setSortId(Double sortId) {
 		this.sortId = sortId;
 	}
